refactor: :recycle: Move ~/utils/configs to ~/configs.
