+++ configs/articleConfig.ts
