+++ configs/articleImageConfig.ts
