+++ configs/authConfig.ts
