+++ configs/errorConfig.ts
