+++ configs/followConfig.ts
