+++ configs/index.ts
