+++ configs/passwordConfig.ts
