+++ configs/reactionConfig.ts
