+++ configs/tagConfig.ts
