+++ configs/viewConfig.ts
