+++ utils/index.ts
@@ -1,5 +1,5 @@
 export * from "./requests";
-export * from "./configs";
+export * from "../configs";
 export * from "./errors";
 export * from "./strings";
 export * from "./schemas";
