+++ utils/requests/articles/indexArticle.ts
@@ -11,7 +11,7 @@ import {
   ArticleCountSchema,
   ArticleAuthSchema,
 } from "~/utils/schemas";
-import { articleConfig } from "~/utils/configs";
+import { articleConfig } from "~/configs";
 
 /* -------------------------------------------------------------------------- */
 /*                             Index article query                            */
